Updates for GrFN loop execution
Update GrFN loop execution to compute the path of the EXIT variable and
check its value before each iteration of the loop body.

+++ automates/program_analysis/CAST2GrFN/model/cast_to_air_model.py
@@ -142,7 +142,7 @@ class C2AVariable(object):
             "mutable": False,  # TODO probably only mutable if object/list/dict type
         }
         if self.type_name == "Number":
-            domain["name"] = "float"
+            domain["name"] = "integer"
         elif self.type_name.startswith("object$"):
             name = self.type_name.split("object$")[-1]
             domain.update(
